fix(tests): rename tests files
