+++ src/Symfony/Component/FeatureFlag/Tests/Provider/InMemoryProviderTest.php
@@ -12,40 +12,28 @@
 namespace Symfony\Component\FeatureFlag\Tests\Provider;
 
 use PHPUnit\Framework\TestCase;
-use Symfony\Component\FeatureFlag\Provider\ChainProvider;
 use Symfony\Component\FeatureFlag\Provider\InMemoryProvider;
 
-class ChainProviderTests extends TestCase
+class InMemoryProviderTest extends TestCase
 {
-    private ChainProvider $provider;
+    private InMemoryProvider $provider;
 
     protected function setUp(): void
     {
-        $this->provider = new ChainProvider([
-            new InMemoryProvider([
-                'first' => fn () => true,
-            ]),
-            new InMemoryProvider([
-                'second' => fn () => 42,
-            ]),
-            new InMemoryProvider([
-                'exception' => fn () => throw new \LogicException('Should not be called.'),
-            ]),
+        $this->provider = new InMemoryProvider([
+            'first' => fn () => true,
+            'second' => fn () => 42,
+            'exception' => fn () => throw new \LogicException('Should not be called.'),
         ]);
     }
 
     public function testGet()
     {
         $feature = $this->provider->get('first');
-
         $this->assertIsCallable($feature);
         $this->assertTrue($feature());
-    }
 
-    public function testGetFallback()
-    {
         $feature = $this->provider->get('second');
-
         $this->assertIsCallable($feature);
         $this->assertSame(42, $feature());
     }
